refactor: extract logic out of index
